Move SampleCollection and lazily import
